Support single-source package version.
Update setup.py to pull __version__ from
tensorflow_similarity/__init__.py using a similar approach to the
setup.py file in the pip package.

https://packaging.python.org/guides/single-sourcing-package-version/

Remove test_requirements.txt and add it to a key in extra_requires. Also
update the README to show how to install a local editble version of the
package with the dev requirements.

Update the update_version.py module to modify the minor version of the
tensorflow_similarity/__init__.py file.

Update github workflows to install the package using pip and the [dev]
key.