Fix init issue in ClassificationMatcher.compile()
Remove tf.constant from default arg value. This was causing TF to
initialize CUDA when the module was imported directly or by any other
module.

The value is now an Optional[FloatTensor] that defaults to None and a
distance threshold of math.inf is used if the value is None.