projector: add notebook renderer
This change introduces Python API to render the new projector in Juypter
notebook. It currently provides work in progress API, `embedding` which
visualizes the embedding data and label rendered in the visualization.

In order to support both Colab and Jupyter, I have decided to introduce
an abstraction, `renderer` which knows how to render an output cell and
how to communicate with it. Do note that Jupyter and Colab are very
different; Jupyter does not encapsulate JavaScript context for each
output cell so a symbol exposed on `globalThis` bleeds over to other
output cells while Colab does not do that.