[bug] Remove reduce_sum from InnerProductSimilarity().
the InnerProductSimilarity() class had a reduce_sum over axis 1. This
was returning an [M, 1] vector containing the sum of the pairwise
similairties for each example. We removed the reduce_sum and added
additionaling test coverage for InnerProductSimilarity() as well as test
for the shape for all the distance classes.

This address the bug raised in issue #214