[fix] nmslib save and load reads and writes from local files first.
nmslib requires a string path and will only read and write to local
files. In order to support writing to a remote file path, we first write
to a local tmp dir and then write that to the user provided path using
tf.io.gfile.GFile.

Also updating the visualization notebook. We re-ran all notebooks to
check the changes and we are adding the updated output of the
visualization notebook to master.