Patch bump to 0.16.7
Bug fix for computing macro avg in retrieval metrics. Previously we took
the unique class ids in the labels, but tf.unique expects a 1D tensor.

Start to migrate tests to use tf.test.TestCase

Remove js projector code for now. Moved to a separate branch.

Thanks to @yonigottesman for:
* converting batchnorm to syncbatchnorm in architecture models to
  support distributed training.
* adding formatting checker for black and isort